Change html.javi and introduce list types
When changing html.javi to use the new list syntax, it was discovered that we needed a list type, so that ADT constructors can be declared with lists as parameters.
This introduced changes to the abstract syntax and parser.

The syntax for declaring a list type was decided to look like this:
type myType = Ctor [Integer]