Refactor patternmatching in Compiler and Interpreter
MatchSingleVal is now a gerenric function residing in the interpreter, which can be used by both Compiler and Interpreter. This makes sure, that when implementing new expression or functionality, we have to fix only one patternmatching.

The tryLookup only needs to be there, because the Variable might be referring to an ADT constructor without parameters, so we need to be able to look up whether the variable refers to an ADT. Not nice, but also don't see how to fix it.